+++ src/main/java/com/dementor/domain/chat/service/ChatRoomService.java
@@ -101,92 +101,28 @@
 
 package com.dementor.domain.chat.service;
 
-import com.dementor.domain.chat.dto.ChatMessageResponseDto;
-import com.dementor.domain.chat.dto.ChatMessageSendDto;
-import com.dementor.domain.chat.dto.ChatMessageSliceDto;
 import com.dementor.domain.chat.dto.ChatRoomResponseDto;
 import com.dementor.domain.chat.entity.ChatMessage;
 import com.dementor.domain.chat.entity.ChatRoom;
-import com.dementor.domain.chat.entity.MessageType;
 import com.dementor.domain.chat.entity.RoomType;
 import com.dementor.domain.chat.repository.ChatMessageRepository;
 import com.dementor.domain.chat.repository.ChatRoomRepository;
-import com.dementor.domain.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.ZoneId;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
-public class ChatService {
+public class ChatRoomService {
 
     private final ChatRoomRepository chatRoomRepository;
     private final ChatMessageRepository chatMessageRepository;
-    private final MemberService memberService; // ✅
+
 
     private static final ZoneId KST = ZoneId.of("Asia/Seoul"); //  시간대 상수화
-
-
-    //메시지 저장 + dto 반환
-    @Transactional
-    public ChatMessageResponseDto handleMessage(ChatMessageSendDto dto, Long memberId, String nickname) {
-        ChatRoom chatRoom = chatRoomRepository.findByApplymentId(dto.getApplymentId())
-                .orElseThrow(() -> new IllegalArgumentException("채팅방이 존재하지 않습니다. applymentId=" + dto.getApplymentId())); // ✅ 예외 메시지 개선
-
-        ChatMessage message = new ChatMessage();
-        message.setChatRoom(chatRoom);
-        message.setType(dto.getType()); // ?메시지 타입 설정 추가
-        message.setNickname(nickname);
-        message.setMemberId(memberId);
-        message.setContent(dto.getType() == MessageType.MESSAGE ? dto.getMessage() : null);
-
-        chatMessageRepository.save(message);
-
-        return new ChatMessageResponseDto(
-                message.getType(),
-                chatRoom.getApplymentId(),
-                memberId,
-                nickname,
-                message.getContent(),
-                message.getCreatedAt().atZone(KST) // 시간대 상수 적용
-        );
-    }
-
-
-
-    // 커서 기반 페이징 + 응답 변환
-    @Transactional(readOnly = true)
-    public ChatMessageSliceDto getMessages(Long chatRoomId, Long beforeMessageId, int size) {
-        ChatRoom chatRoom = chatRoomRepository.findById(chatRoomId)
-                .orElseThrow(() -> new IllegalArgumentException("채팅방이 존재하지 않습니다. chatRoomId=" + chatRoomId)); // 예외 메시지 개선
-
-        List<ChatMessage> messages;
-
-        if (beforeMessageId != null) {
-            messages = chatMessageRepository.findTop20ByChatRoom_ChatRoomIdAndChatMessageIdLessThanOrderByChatMessageIdDesc(chatRoomId, beforeMessageId);
-        } else {
-            messages = chatMessageRepository.findTop20ByChatRoom_ChatRoomIdOrderByChatMessageIdDesc(chatRoomId); // 최신 메시지 size개 조회
-        }
-
-        List<ChatMessageResponseDto> dtoList = messages.stream()
-                .map(m -> new ChatMessageResponseDto(
-                        m.getType(),
-                        m.getChatRoom().getApplymentId(),
-                        m.getMemberId(),
-                        m.getNickname(),
-                        m.getContent(),
-                        m.getCreatedAt().atZone(KST) // 시간대 상수 적용
-                ))
-                .collect(Collectors.toList());
-
-        Long nextCursor = messages.isEmpty() ? null : messages.get(messages.size() - 1).getChatMessageId();
-
-        return new ChatMessageSliceDto(dtoList, dtoList.size() == size, nextCursor);
-    }
